# [WIP] Optimize bash and python