--- a/bash/wordcount.sh
+++ b/bash/wordcount.sh
@@ -1,3 +1,4 @@
 #!/usr/bin/env bash
 export LC_COLLATE=C
-sed 's/[\t ]/\n/g' | grep -v ^$ | sort | uniq -c | sed 's/^\s*//' | sort  -k1,2nr -k2 | awk 'BEGIN{OFS="\t"}{print $2,$1}'
+export PHYS_CORES=$(cat /proc/cpuinfo | grep 'core id' | sort | uniq | wc -l)
+sed -E 's/[\t ]+/\n/g' | grep -v ^$ | sort -S 40% --parallel $PHYS_CORES | uniq -c | sort -S 40% ---parallel $PHYS_CORES 2 -k1,2nr -k2 | awk 'BEGIN{OFS="\t"}{print $2,$1}'