Replaces https://github.com/juditacs/wordcount/pull/72 since similar changes have gotten pulled into 
other PRs and I don't want merge conflicts.

Benchmark: `time cat data/huwikisource-latest-pages-meta-current.xml | mycommand > /dev/null`
System: Macbook Pro mid-2014 edition, 2.2 GHz Core i7
Times: real / CPU/ sys

Bash optimization history:
-  TODO show benchmarks

Python optimizations:

0. Baseline:  25.631s / 25.145s / 0.493s - output md5sum b52070c4fdd3b0b87ea5659e2e862128 
1. Remove UTF-8 encode/decode, and directly sort list:14.633s / 14.302s / 0.368s - output md5sum: b52070c4fdd3b0b87ea5659e2e862128
2. Partition and sort 

Baseline: 25-26s to run on 

- [ ]